Move mouse cursor functionality outside of VBL ISR
